+++ layers/Schema/packages/schema-commons/src/FilterInputProcessors/FilterInputProcessor.php
@@ -14,6 +14,7 @@ class FilterInputProcessor extends AbstractFilterInputProcessor
     public const FILTERINPUT_SEARCH = 'filterinput-search';
     public const FILTERINPUT_DATES = 'filterinput-dates';
     public const FILTERINPUT_INCLUDE = 'filterinput-include';
+    public const FILTERINPUT_EXCLUDE_IDS = 'filterinput-exclude-ids';
     public const FILTERINPUT_PARENT_IDS = 'filterinput-parent-ids';
     public const FILTERINPUT_PARENT_ID = 'filterinput-parent-id';
 
@@ -26,6 +27,7 @@ class FilterInputProcessor extends AbstractFilterInputProcessor
             [self::class, self::FILTERINPUT_SEARCH],
             [self::class, self::FILTERINPUT_DATES],
             [self::class, self::FILTERINPUT_INCLUDE],
+            [self::class, self::FILTERINPUT_EXCLUDE_IDS],
             [self::class, self::FILTERINPUT_PARENT_IDS],
             [self::class, self::FILTERINPUT_PARENT_ID],
         );
@@ -53,6 +55,9 @@ class FilterInputProcessor extends AbstractFilterInputProcessor
                 break;
             case self::FILTERINPUT_INCLUDE:
                 $query['include'] = $value;
+                break;
+            case self::FILTERINPUT_EXCLUDE_IDS:
+                $query['exclude-ids'] = $value;
                 break;
             case self::FILTERINPUT_PARENT_IDS:
                 $query['parent-ids'] = $value;