+++ layers/Schema/packages/users-wp/src/TypeAPIs/UserTypeAPI.php
@@ -167,9 +167,9 @@ class UserTypeAPI extends AbstractUserTypeAPI
             // Transform from array to string
             $query['include'] = implode(',', $query['include']);
         }
-        if (isset($query['exclude'])) {
-            // Transform from array to string
-            $query['exclude'] = implode(',', $query['exclude']);
+        if (isset($query['exclude-ids'])) {
+            $query['exclude'] = $query['exclude-ids'];
+            unset($query['exclude-ids']);
         }
         if (isset($query['order'])) {
             // Same param name, so do nothing